% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{castvote_cache_dir}
\alias{castvote_cache_dir}
\title{Cache directory}
\usage{
castvote_cache_dir()
}
\value{
A path.
}
\description{
Where downloaded source files are stored. Override with the
`castvote.cache_dir` option.
}
