% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{sf_rules}
\alias{sf_rules}
\title{San Francisco tabulation rules}
\usage{
sf_rules(year)
}
\arguments{
\item{year}{Election year. Only 2018 is defined.}
}
\value{
An `irv_rules` object.
}
\description{
San Francisco tabulation rules
}
